Добавил сброс формы
