+++ js/util.js
@@ -1,6 +1,3 @@
-const ALERT_SHOW_TIME = 5000;
-const main = document.querySelector('main');
-
 const getRandomNumber = (min, max, digit = 0) => {
   if (min < 0 || max < 0) {
     return -1;
@@ -39,45 +36,4 @@ const numDecline = (num, nominative, genitiveSingular, genitivePlural) => {
   }
 };
 
-const showAlert = (message) => {
-  const alertContainer = document.createElement('div');
-  alertContainer.style.zIndex = 100;
-  alertContainer.style.position = 'absolute';
-  alertContainer.style.left = 0;
-  alertContainer.style.top = 0;
-  alertContainer.style.right = 0;
-  alertContainer.style.padding = '10px 3px';
-  alertContainer.style.fontSize = '20px';
-  alertContainer.style.textAlign = 'center';
-  alertContainer.style.backgroundColor = 'red';
-
-  alertContainer.textContent = message;
-
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, ALERT_SHOW_TIME);
-}
-
-const showSuccessMessage = () => {
-  const successMessageTemplate = document.querySelector('#success').content;
-  const successMessage = successMessageTemplate.querySelector('.success');
-  main.appendChild(successMessage);
-
-  setTimeout(() => {
-    successMessage.remove();
-  }, ALERT_SHOW_TIME);
-};
-
-const showErrorMessage = () => {
-  const errorMessageTemplate = document.querySelector('#error').content;
-  const errorMessage = errorMessageTemplate.querySelector('.error');
-  main.appendChild(errorMessage);
-
-  setTimeout(() => {
-    errorMessage.remove();
-  }, ALERT_SHOW_TIME);
-};
-
-export {getRandomNumber, getRandomArrayElement, shuffleArray, numDecline, showAlert, showSuccessMessage, showErrorMessage};
+export {getRandomNumber, getRandomArrayElement, shuffleArray, numDecline};
